# sync/openapi.yaml
...
openapi: 3.1.0
info:
  title: YunPin encrypted sync relay API
  version: 1.0.0-dev
  description: >-
    A content-blind relay for end-to-end encrypted YunPin state. All byte fields
    use unpadded base64url unless explicitly documented as 32-character hex IDs.
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - deviceBearer: []
paths:
...
  /v1/auth/register:
    post:
      operationId: registerUser
      security: []
      description: >-
        Creates a local self-hosted YunPin login. The relay stores a salted
        PBKDF2-SHA-256 password verifier and only a hash of the returned
        opaque session token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserPassword"
      responses:
        "201":
          $ref: "#/components/responses/UserSession"
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "429":
          $ref: "#/components/responses/Error"
  /v1/auth/login:
    post:
      operationId: loginUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserPassword"
      responses:
        "200":
          $ref: "#/components/responses/UserSession"
        "401":
          $ref: "#/components/responses/Error"
        "429":
          $ref: "#/components/responses/Error"
  /v1/auth/logout:
    post:
      operationId: logoutUser
      security:
        - userSession: []
      responses:
        "204":
          description: Session revoked.
        "401":
          $ref: "#/components/responses/Error"
  /v1/accounts/{accountId}/claim:
    parameters:
      - $ref: "#/components/parameters/AccountId"
    post:
      operationId: claimAccount
      security:
        - userSession: []
      description: >-
        Binds an existing account to the logged-in user. The request must
        prove possession of the account's current active device credential,
        which is already protected by the local OS credential store. It does
        not accept a recovery key, password, or recovery-derived value.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [device_id, device_token]
              properties:
                device_id:
                  $ref: "#/components/schemas/ID"
                device_token:
                  $ref: "#/components/schemas/SecretInput"
      responses:
        "200":
          description: Account is now bound to the current user.
        "401":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    deviceBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque-device-token
    userSession:
      type: http
      scheme: bearer
      bearerFormat: opaque-user-session
  parameters:
    AccountId:
      name: accountId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/HexId"
...
  responses:
    Error:
      description: Stable machine-readable error code.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    UserSession:
      description: Opaque login session. Store the token only in the platform secret store.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/UserSession"
  schemas:
    HexId:
      type: string
      pattern: "^[0-9a-f]{32}$"
      examples: ["00112233445566778899aabbccddeeff"]
    Base64Url:
      type: string
      pattern: "^[A-Za-z0-9_-]+$"
    SecretInput:
      allOf:
        - $ref: "#/components/schemas/Base64Url"
      writeOnly: true
    UserPassword:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username:
          type: string
          pattern: "^[A-Za-z0-9._-]{3,64}$"
        password:
          type: string
          format: password
          writeOnly: true
          minLength: 12
          maxLength: 256
    UserSession:
      type: object
      additionalProperties: false
      required: [username, token, expires_at]
      properties:
        username:
          type: string
        token:
          allOf:
            - $ref: "#/components/schemas/OneTimeSecret"
          description: Opaque 32-byte login session; never include it in endpoint configuration.
        expires_at:
          type: string
          format: date-time
    OneTimeSecret:
      allOf:
        - $ref: "#/components/schemas/Base64Url"
      readOnly: true
...
    Error:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: string